requirements.txt
tqdm
h5py
torch
numpy
typer
gdown
kaggle
pandas
Pillow
fastdl
optuna
plotly
joblib
pydicom
seaborn
streamlit
onnx2torch
matplotlib
torchvision
torchmetrics
scikit-learn
python-dotenv
pytorch-lightning
